Asda bug fixed.
On asda web page, the button was only clickable when it is appeared on the screen somehow. I added codes to bring the screen around the button area now it works. Also stopped checking cookies for every single loop, instead when driver starts, It checks only there and thats it.