Asda added
It was so stressful omg. The litre was written as lt,, 10g was present idk why, 75cl was written as 75c3 and the pence was written in float and there was big typo which was one of these cheeses price was written as 12,000.00 / kg which was supposed to be 12.00. nearly doomed my analyses. Anyways I fixed it haha! Good stuff